Radiation Changes
Radioactivity overhauled:

- players have permanent contamination.
- players have internal damage

as contamination increases, permanent contamination increases by a factor of ~0.000001

permanent contamination slowly causes players to increase their internal damage.

as internal damage increases, players will suffer the effects of radiation poisoning just like previous versions, with an added effect of losing max health to go with it.

at a constant exposure of 14.4 Rad/s it would take a player >3.5 in-game days to succumb to radiation poisoning.

IMPORTANT: When implementing max health related changes, update entity capability: Type.OLD_MAX_HEALTH. (20 + whatever modifications) p.s. this makes the mod incompatible with other mods that add max health, add compatibilities later

+++ src/main/java/com/hbm/nucleartech/item/custom/GeigerCounterItem.java
@@ -129,8 +129,8 @@ public class GeigerCounterItem extends Item {
 
     public static float getRad(LivingEntity pEntity) {
 
-        return HbmCapabilities.getData(pEntity).getValue(IEntityCapabilityBase.Type.RADENV)
-                + HbmCapabilities.getData(pEntity).getValue(IEntityCapabilityBase.Type.NEUTRON);
+        return (float)(HbmCapabilities.getData(pEntity).getValue(IEntityCapabilityBase.Type.RADENV)
+                + HbmCapabilities.getData(pEntity).getValue(IEntityCapabilityBase.Type.NEUTRON));
     }
 
     @Override
@@ -188,7 +188,7 @@ public class GeigerCounterItem extends Item {
                     if(System.currentTimeMillis() >= lastRadSurvey + 1000) {
                         lastRadSurvey = System.currentTimeMillis();
                         radPS = getRad(mc.player);
-                        radiation = HbmCapabilities.getData(mc.player).getValue(IEntityCapabilityBase.Type.RADIATION);
+                        radiation = (float)HbmCapabilities.getData(mc.player).getValue(IEntityCapabilityBase.Type.RADIATION);
                     }
 
 //                    int screenWidth = mc.getWindow().getGuiScaledWidth();
